requirements.txt
# AaryaOnlineCompiler - Python Dependencies
# Created by Aarya Agarwal

Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
asgiref==3.8.1
sqlparse==0.5.3
